renamed DigitizeUtil to AnimateUtil
