+++ src/Util/AnimateUtil/AnimateUtil.spec.js
@@ -3,14 +3,14 @@ import OlGeomPoint from 'ol/geom/point';
 import OlFeature from 'ol/feature';
 import OlStyleStyle from 'ol/style/style';
 
-import DigitizeUtil from '../DigitizeUtil/DigitizeUtil';
+import AnimateUtil from '../AnimateUtil/AnimateUtil';
 import TestUtil from '../TestUtil';
 
-describe('DigitizeUtil', () => {
+describe('AnimateUtil', () => {
 
   describe('Basics', () => {
     it('is defined', () => {
-      expect(DigitizeUtil).toBeDefined();
+      expect(AnimateUtil).toBeDefined();
     });
   });
 
@@ -18,25 +18,21 @@ describe('DigitizeUtil', () => {
 
     describe('#moveFeature', () => {
       it('is defined', () => {
-        expect(DigitizeUtil.moveFeature).toBeDefined();
+        expect(AnimateUtil.moveFeature).toBeDefined();
       });
       it('registers postcompose listener on the map', () => {
 
-        const geom = new OlGeomPoint([0, 0]);
+        const coords = [0, 0];
+        const geom = new OlGeomPoint(coords);
         const featToMove = new OlFeature(geom);
 
         let map = TestUtil.createMap();
 
-        expect(map.listeners_.postcompose).toBeUndefined();
-
-        const expectedKey = DigitizeUtil.moveFeature(
+        AnimateUtil.moveFeature(
           map, featToMove, 100, 50, new OlStyleStyle()
-        );
-
-        expect(typeof map.listeners_.postcompose).toBe('object');
-        expect(typeof expectedKey).toBe('object');
-        expect(expectedKey.type).toBe('postcompose');
-
+        ).then((feat) => {
+          expect(feat.getGeometry().getCoordinates()).toEqual([50, 50]);
+        });
         TestUtil.removeMap(map);
       });
     });
