+++ src/index.js
@@ -25,7 +25,7 @@ import Window from './Window/Window.jsx';
 
 import CapabilitiesUtil from './Util/CapabilitiesUtil/CapabilitiesUtil';
 import CsrfUtil from './Util/CsrfUtil/CsrfUtil';
-import DigitizeUtil from './Util/DigitizeUtil/DigitizeUtil';
+import AnimateUtil from './Util/AnimateUtil/AnimateUtil';
 import FeatureUtil from './Util/FeatureUtil/FeatureUtil';
 import MapUtil from './Util/MapUtil/MapUtil';
 import MeasureUtil from './Util/MeasureUtil/MeasureUtil';
@@ -64,7 +64,7 @@ export {
   Window,
   CapabilitiesUtil,
   CsrfUtil,
-  DigitizeUtil,
+  AnimateUtil,
   FeatureUtil,
   Logger,
   MapUtil,
